Split compare page overrides by locale
