+++ frontend/app/(localized)/[locale]/(marketing)/ai-video-engines/[slug]/_lib/compare-page-overrides.ts
@@ -1,640 +1,14 @@
 import type { AppLocale } from '@/i18n/locales';
+import { EN_COMPARE_PAGE_OVERRIDES } from './compare-page-overrides-en';
+import { ES_COMPARE_PAGE_OVERRIDES } from './compare-page-overrides-es';
+import { FR_COMPARE_PAGE_OVERRIDES } from './compare-page-overrides-fr';
+import type { ComparePageOverride, ComparePageOverridesByLocale } from './compare-page-overrides-types';
+export type { ComparePageOverride } from './compare-page-overrides-types';
 
-export type ComparePageOverride = {
-  meta?: {
-    title?: string;
-    description?: string;
-  };
-  heroIntro?: string;
-  topCards?: Array<{
-    title: string;
-    body: string;
-  }>;
-  primaryLinksTitle?: string;
-  primaryLinks?: Array<{
-    href: string;
-    label: string;
-  }>;
-  faq?: {
-    title?: string;
-    subtitle?: string;
-    items: Array<{
-      question: string;
-      answer: string | string[];
-    }>;
-  };
-};
-
-const COMPARE_PAGE_OVERRIDES: Partial<Record<AppLocale, Record<string, ComparePageOverride>>> = {
-  en: {
-    'seedance-1-5-pro-vs-seedance-2-0': {
-      meta: {
-        title: 'Seedance 1.5 Pro vs Seedance 2.0 | What Changed, Upgrade Path & Best Use Cases | MaxVideoAI',
-        description:
-          'Compare Seedance 1.5 Pro vs Seedance 2.0 on MaxVideoAI to see what changed in audio, multi-shot continuity, references, pricing, and when upgrading makes sense.',
-      },
-      heroIntro:
-        'Compare Seedance 1.5 Pro and Seedance 2.0 to see what changed between the older Seedance Pro workflow and the current Seedance AI video model in native audio, multi-shot continuity, and reference workflows. Use this page to understand the trade-offs quickly before moving to the current Seedance model, the Seedance AI video examples page, or the exact Seedance video workflow that fits your use case.',
-      topCards: [
-        {
-          title: 'What changed',
-          body:
-            'Seedance 2.0 is the newer Seedance AI video workflow with stronger multi-shot continuity, broader reference inputs, and a more current audio-first production path than Seedance 1.5 Pro.',
-        },
-        {
-          title: 'When to stay on Seedance 1.5 Pro',
-          body:
-            'Stay on Seedance 1.5 Pro when you mainly need short, repeatable clips, simpler camera setups, and an older workflow that is already validated in production.',
-        },
-        {
-          title: 'When to upgrade to Seedance 2.0',
-          body:
-            'Upgrade when you need better shot-to-shot continuity, richer native audio workflows, or a more flexible current model for higher-value creative work.',
-        },
-        {
-          title: 'Best use cases',
-          body:
-            'Use this page to decide between a supported older Seedance workflow for short controlled clips and the current Seedance 2.0 workflow for multi-shot ads, launches, and more ambitious reference-driven sequences.',
-        },
-      ],
-      primaryLinksTitle: 'Recommended next steps',
-      primaryLinks: [
-        {
-          href: '/models/seedance-2-0',
-          label: 'Open Seedance 2.0 model page',
-        },
-        {
-          href: '/examples/seedance',
-          label: 'View Seedance AI video examples',
-        },
-        {
-          href: '/models/seedance-2-0-fast',
-          label: 'Open Seedance 2.0 Fast model page',
-        },
-        {
-          href: '/ai-video-engines/seedance-2-0-vs-seedance-2-0-fast',
-          label: 'Compare Seedance 2.0 vs Seedance 2.0 Fast',
-        },
-      ],
-      faq: {
-        title: 'FAQ',
-        subtitle:
-          'Short answers to help you decide whether to stay on Seedance 1.5 Pro or move to Seedance 2.0.',
-        items: [
-          {
-            question: 'What changed between Seedance 1.5 Pro and Seedance 2.0?',
-            answer:
-              'Seedance 2.0 is the newer model with stronger multi-shot continuity, broader reference workflows, and the current Seedance path for production work. Seedance 1.5 Pro remains useful for shorter, simpler, repeatable clips.',
-          },
-          {
-            question: 'Is Seedance 2.0 better than Seedance 1.5 Pro?',
-            answer:
-              'For most current workflows, yes. Seedance 2.0 is the better default if you want the current Seedance AI video model for continuity, flexibility, and broader production use, while Seedance 1.5 Pro remains useful as an older Seedance Pro setup for shorter clips.',
-          },
-          {
-            question: 'When should I upgrade from Seedance 1.5 Pro to Seedance 2.0?',
-            answer:
-              'Upgrade when you need better multi-shot behavior, richer native audio workflows, or more headroom for current prompt and reference-driven production. If your existing 1.5 Pro workflow is already stable for short clips, you do not need to move every job immediately.',
-          },
-          {
-            question: 'Is Seedance 1.5 Pro still good enough for some workflows?',
-            answer:
-              'Yes. It still fits short, repeatable cinematic clips and teams that already have validated prompt patterns on 1.5 Pro and do not need the broader 2.0 workflow yet.',
-          },
-          {
-            question: 'Which model is better for multi-shot and native audio?',
-            answer:
-              'Seedance 2.0 is the better choice for multi-shot continuity and the more current native-audio workflow. Seedance 1.5 Pro is better treated as a simpler older option for shorter clips.',
-          },
-        ],
-      },
-    },
-    'seedance-2-0-vs-seedance-2-0-fast': {
-      heroIntro:
-        'Compare Seedance 2.0 and Seedance 2.0 Fast to choose the right current Seedance AI video workflow for final multi-shot work, native audio, and workflow comparison. Use this page to see when standard Seedance is better for polished Seedance video output and when Fast is better for testing, timing checks, and cheaper iteration.',
-      primaryLinksTitle: 'Recommended next steps',
-      primaryLinks: [
-        {
-          href: '/models/seedance-2-0',
-          label: 'Open Seedance 2.0 model page',
-        },
-        {
-          href: '/examples/seedance',
-          label: 'View Seedance AI video examples',
-        },
-        {
-          href: '/models/seedance-2-0-fast',
-          label: 'Open Seedance 2.0 Fast model page',
-        },
-      ],
-      faq: {
-        title: 'FAQ',
-        subtitle:
-          'Short answers to help you choose the right current Seedance workflow.',
-        items: [
-          {
-            question: 'Which Seedance AI video model should I use for draft work?',
-            answer:
-              'Use Seedance 2.0 Fast for cheaper draft passes, faster testing, and workflow comparisons. Use Seedance 2.0 when you want stronger final-quality multi-shot output, native audio, and a more production-ready workflow.',
-          },
-          {
-            question: 'How is Seedance 2.0 different from Seedance 2.0 Fast?',
-            answer:
-              'Seedance 2.0 is the stronger current choice for polished multi-shot work, native audio, and more demanding reference-driven outputs, while Seedance 2.0 Fast is better for cheaper tests, timing checks, and early iteration.',
-          },
-          {
-            question: 'Is Seedance 2.0 better for polished Seedance video output?',
-            answer:
-              'Yes. Seedance 2.0 is the better fit when the goal is polished Seedance video output, while Fast is the better fit when the goal is to test ideas and compare workflows quickly.',
-          },
-        ],
-      },
-    },
-    'veo-3-1-vs-veo-3-1-fast': {
-      heroIntro:
-        'Compare Veo 3.1 and Veo 3.1 Fast to choose the right current Veo 3 AI workflow for polished text-to-video, image-to-video, faster draft passes, and native-audio control.',
-      faq: {
-        title: 'FAQ',
-        subtitle:
-          'Short answers to help you choose the right current Veo workflow.',
-        items: [
-          {
-            question: 'How should I use Veo 3 for text-to-video and draft testing?',
-            answer:
-              'Use Veo 3.1 Fast for cheaper draft passes, text-to-video prompt comparison, and quicker iteration. Use Veo 3.1 when you want stronger final-quality output, richer reference-guided control, and more polished image-to-video results.',
-          },
-          {
-            question: 'Can I use both Veo 3.1 and Veo 3.1 Fast for image-to-video?',
-            answer:
-              'Yes. Both can handle image-to-video workflows, but Veo 3.1 is the better fit for more polished results while Veo 3.1 Fast is the better fit for cheaper prompt and framing tests.',
-          },
-          {
-            question: 'When should I choose Veo 3.1 instead of Veo 3.1 Fast?',
-            answer:
-              'Choose Veo 3.1 when final quality, native audio polish, and stronger reference-guided control matter more than draft speed. Choose Fast when the goal is cheaper iteration and quicker workflow validation.',
-          },
-        ],
-      },
-    },
-    'veo-3-1-fast-vs-veo-3-1-lite': {
-      heroIntro:
-        'Compare Veo 3.1 Fast and Veo 3.1 Lite to choose the right current Veo 3 AI workflow for cheaper text-to-video tests, image-to-video tests, native-audio behavior, and faster iteration.',
-      faq: {
-        title: 'FAQ',
-        subtitle:
-          'Short answers to help you choose between the current Fast and Lite Veo tiers.',
-        items: [
-          {
-            question: 'Which Veo 3 model is better for image-to-video tests?',
-            answer:
-              'Both can work, but Veo 3.1 Lite is better for the cheapest audio-ready image-to-video tests, while Veo 3.1 Fast is better when you want broader flexibility and a smoother upgrade path into full Veo 3.1.',
-          },
-          {
-            question: 'Is Veo 3.1 Lite or Veo 3.1 Fast better for text-to-video drafts?',
-            answer:
-              'Veo 3.1 Lite is better when you want the lowest-cost audio-ready tests. Veo 3.1 Fast is better when you want more output flexibility, optional audio, and a cleaner bridge into the main Veo 3.1 workflow.',
-          },
-          {
-            question: 'When should I choose Veo 3.1 Fast instead of Veo 3.1 Lite?',
-            answer:
-              'Choose Fast when you want broader workflow flexibility, optional audio control, and an easier upgrade path into Veo 3.1. Choose Lite when your priority is the cheapest current Veo testing with audio always on.',
-          },
-        ],
-      },
-    },
-    'kling-3-pro-vs-kling-3-standard': {
-      heroIntro:
-        'Compare Kling 3 Pro and Kling 3 Standard to choose the right current Kling AI model for multi-shot video, Kling image-to-video workflows, reusable Kling Elements, and native-audio output quality.',
-      faq: {
-        title: 'FAQ',
-        subtitle:
-          'Short answers to help you choose between the current Kling Pro and Standard tiers.',
-        items: [
-          {
-            question: 'Which current Kling AI model is better for image-to-video and prompt testing?',
-            answer:
-              'Kling 3 Standard is better for lower-cost prompt testing and repeatable image-to-video tests, while Kling 3 Pro is better when you need tighter scene control and higher-priority final outputs.',
-          },
-          {
-            question: 'Do Kling 3 Pro and Kling 3 Standard both support Kling Elements?',
-            answer:
-              'Yes. Both current Kling models support Kling Elements for character and prop continuity, but Kling 3 Pro is the stronger choice when the sequence is more demanding or continuity matters more.',
-          },
-          {
-            question: 'When should I choose Kling 3 Pro instead of Kling 3 Standard?',
-            answer:
-              'Choose Kling 3 Pro when you need stronger scene control, more demanding multi-shot continuity, and higher-priority final output. Choose Kling 3 Standard when cost control and repeatable draft testing matter more.',
-          },
-        ],
-      },
-    },
-  },
-  fr: {
-    'seedance-1-5-pro-vs-seedance-2-0': {
-      meta: {
-        title: 'Seedance 1.5 Pro vs Seedance 2.0 | Ce qui change, quand passer a 2.0 et meilleurs cas d usage | MaxVideoAI',
-        description:
-          'Comparez Seedance 1.5 Pro et Seedance 2.0 sur MaxVideoAI pour voir ce qui change sur l audio, la continuite multi-shot, les references, le prix, et quand passer a 2.0.',
-      },
-      heroIntro:
-        'Comparez Seedance 1.5 Pro et Seedance 2.0 pour voir ce qui change entre l ancien workflow Seedance Pro et le modele video IA Seedance actuel sur l audio natif, la continuite multi-shot et les workflows a references. Utilisez cette page pour comprendre rapidement les compromis avant d ouvrir le modele Seedance actuel, la page d exemples video IA Seedance, ou le workflow video Seedance le plus adapte a votre usage.',
-      topCards: [
-        {
-          title: 'Ce qui change',
-          body:
-            'Seedance 2.0 est le workflow video IA Seedance le plus recent, avec une meilleure continuite multi-shot, des entrees de reference plus larges, et une approche audio-first plus actuelle que Seedance 1.5 Pro.',
-        },
-        {
-          title: 'Quand rester sur Seedance 1.5 Pro',
-          body:
-            'Restez sur Seedance 1.5 Pro si vous avez surtout besoin de clips courts et repetables, de setups camera plus simples, et d un workflow plus ancien deja valide en production.',
-        },
-        {
-          title: 'Quand passer a Seedance 2.0',
-          body:
-            'Passez a Seedance 2.0 si vous avez besoin d une meilleure continuite entre plans, de workflows audio natifs plus riches, ou d un modele actuel plus flexible pour un travail creatif a plus forte valeur.',
-        },
-        {
-          title: 'Meilleurs cas d usage',
-          body:
-            'Cette page sert a choisir entre un workflow Seedance plus ancien mais encore supporte pour des clips courts et controles, et le workflow Seedance 2.0 actuel pour des pubs multi-shot, des lancements, et des sequences plus ambitieuses guidees par references.',
-        },
-      ],
-      primaryLinksTitle: 'Parcours recommande',
-      primaryLinks: [
-        {
-          href: '/models/seedance-2-0',
-          label: 'Ouvrir la page modele Seedance 2.0',
-        },
-        {
-          href: '/examples/seedance',
-          label: 'Voir les exemples video IA Seedance',
-        },
-        {
-          href: '/models/seedance-2-0-fast',
-          label: 'Ouvrir la page modele Seedance 2.0 Fast',
-        },
-        {
-          href: '/ai-video-engines/seedance-2-0-vs-seedance-2-0-fast',
-          label: 'Comparer Seedance 2.0 vs Seedance 2.0 Fast',
-        },
-      ],
-      faq: {
-        title: 'FAQ',
-        subtitle:
-          'Reponses rapides pour decider s il faut rester sur Seedance 1.5 Pro ou passer a Seedance 2.0.',
-        items: [
-          {
-            question: 'Qu est-ce qui change entre Seedance 1.5 Pro et Seedance 2.0 ?',
-            answer:
-              'Seedance 2.0 est le modele le plus recent, avec une meilleure continuite multi-shot, des workflows de reference plus larges, et le chemin Seedance actuel pour la production. Seedance 1.5 Pro reste utile pour des clips plus courts, plus simples et repetables.',
-          },
-          {
-            question: 'Seedance 2.0 est-il meilleur que Seedance 1.5 Pro ?',
-            answer:
-              'Pour la plupart des workflows actuels, oui. Seedance 2.0 est le meilleur choix par defaut si vous cherchez le modele video IA Seedance actuel pour plus de continuite, plus de flexibilite et un usage de production plus large, tandis que Seedance 1.5 Pro reste utile comme setup Seedance Pro plus ancien pour des clips courts.',
-          },
-          {
-            question: 'Quand faut-il passer de Seedance 1.5 Pro a Seedance 2.0 ?',
-            answer:
-              'Passez a 2.0 si vous avez besoin d un meilleur comportement multi-shot, de workflows audio natifs plus riches, ou de plus de marge pour des productions actuelles basees sur prompts et references. Si votre workflow 1.5 Pro est deja stable sur des clips courts, vous n avez pas besoin de migrer tous les jobs tout de suite.',
-          },
-          {
-            question: 'Seedance 1.5 Pro est-il encore suffisant pour certains workflows ?',
-            answer:
-              'Oui. Il reste adapte a des clips cinematographiques courts et repetables, ainsi qu aux equipes qui ont deja des prompts valides sur 1.5 Pro et n ont pas encore besoin du workflow 2.0 plus large.',
-          },
-          {
-            question: 'Quel modele est le meilleur pour le multi-shot et l audio natif ?',
-            answer:
-              'Seedance 2.0 est le meilleur choix pour la continuite multi-shot et pour le workflow audio natif le plus actuel. Seedance 1.5 Pro doit plutot etre traite comme une option plus simple et plus ancienne pour des clips courts.',
-          },
-        ],
-      },
-    },
-    'seedance-2-0-vs-seedance-2-0-fast': {
-      heroIntro:
-        'Comparez Seedance 2.0 et Seedance 2.0 Fast pour choisir le bon workflow video IA Seedance actuel selon votre besoin de rendu multi-shot final, d audio natif et de comparaison de workflow. Utilisez cette page pour voir quand le Seedance standard convient mieux a une sortie video Seedance soignee et quand Fast convient mieux aux tests, aux checks de timing et a l iteration moins couteuse.',
-      primaryLinksTitle: 'Parcours recommande',
-      primaryLinks: [
-        {
-          href: '/models/seedance-2-0',
-          label: 'Ouvrir la page modele Seedance 2.0',
-        },
-        {
-          href: '/examples/seedance',
-          label: 'Voir les exemples video IA Seedance',
-        },
-        {
-          href: '/models/seedance-2-0-fast',
-          label: 'Ouvrir la page modele Seedance 2.0 Fast',
-        },
-      ],
-      faq: {
-        title: 'FAQ',
-        subtitle:
-          'Reponses courtes pour choisir le bon workflow Seedance actuel.',
-        items: [
-          {
-            question: 'Quel modele video IA Seedance faut-il utiliser pour le travail de draft ?',
-            answer:
-              'Utilisez Seedance 2.0 Fast pour des tests moins couteux, des tests plus rapides et des comparaisons de workflow. Utilisez Seedance 2.0 quand vous voulez une sortie multi-shot plus aboutie, l audio natif et un workflow plus pret pour la production.',
-          },
-          {
-            question: 'Quelle difference entre Seedance 2.0 et Seedance 2.0 Fast ?',
-            answer:
-              'Seedance 2.0 est le choix actuel le plus solide pour un rendu multi-shot soigne, l audio natif et des sorties a references plus exigeantes, tandis que Seedance 2.0 Fast convient mieux aux tests moins couteux, aux checks de timing et a l iteration initiale.',
-          },
-          {
-            question: 'Seedance 2.0 est-il meilleur pour une sortie video Seedance soignee ?',
-            answer:
-              'Oui. Seedance 2.0 convient mieux quand l objectif est une sortie video Seedance soignee, tandis que Fast convient mieux quand l objectif est de tester des idees et de comparer rapidement des workflows.',
-          },
-        ],
-      },
-    },
-    'veo-3-1-vs-veo-3-1-fast': {
-      heroIntro:
-        'Comparez Veo 3.1 et Veo 3.1 Fast pour choisir le bon workflow video IA Veo 3 actuel selon votre besoin de texte-to-video soigne, d image-to-video, de tests plus rapides et de controle sur l audio natif.',
-      faq: {
-        title: 'FAQ',
-        subtitle:
-          'Reponses courtes pour choisir le bon workflow Veo actuel.',
-        items: [
-          {
-            question: 'Comment utiliser Veo 3 pour le texte-to-video et les tests ?',
-            answer:
-              'Utilisez Veo 3.1 Fast pour des tests moins chers, la comparaison de prompts texte-to-video et une iteration plus rapide. Utilisez Veo 3.1 quand vous voulez une sortie finale plus solide, un meilleur controle guide par references et des resultats image-to-video plus soignes.',
-          },
-          {
-            question: 'Peut-on utiliser Veo 3.1 et Veo 3.1 Fast pour l image-to-video ?',
-            answer:
-              'Oui. Les deux peuvent gerer des workflows image-to-video, mais Veo 3.1 convient mieux a des resultats plus aboutis, tandis que Veo 3.1 Fast convient mieux a des tests de prompt et de cadrage moins chers.',
-          },
-          {
-            question: 'Quand faut-il choisir Veo 3.1 plutot que Veo 3.1 Fast ?',
-            answer:
-              'Choisissez Veo 3.1 quand la qualite finale, la finition sur l audio natif et un meilleur controle guide par references comptent plus que la vitesse de draft. Choisissez Fast quand l objectif est une iteration moins chere et une validation de workflow plus rapide.',
-          },
-        ],
-      },
-    },
-    'veo-3-1-fast-vs-veo-3-1-lite': {
-      heroIntro:
-        'Comparez Veo 3.1 Fast et Veo 3.1 Lite pour choisir le bon workflow video IA Veo 3 actuel selon votre besoin de tests texte-to-video moins chers, de tests image-to-video, de comportement audio et d iteration plus rapide.',
-      faq: {
-        title: 'FAQ',
-        subtitle:
-          'Reponses courtes pour choisir entre les tiers Veo Fast et Lite actuels.',
-        items: [
-          {
-            question: 'Quel modele Veo 3 convient le mieux pour des tests image-to-video ?',
-            answer:
-              'Les deux peuvent convenir, mais Veo 3.1 Lite est plus adapte aux tests image-to-video audio-ready les moins chers, tandis que Veo 3.1 Fast est plus adapte quand vous voulez plus de flexibilite et une montee plus fluide vers Veo 3.1.',
-          },
-          {
-            question: 'Veo 3.1 Lite ou Veo 3.1 Fast convient mieux a des tests texte-to-video ?',
-            answer:
-              'Veo 3.1 Lite convient mieux quand vous voulez les tests audio-ready les moins chers. Veo 3.1 Fast convient mieux quand vous voulez plus de flexibilite, un audio optionnel et un pont plus propre vers le workflow principal Veo 3.1.',
-          },
-          {
-            question: 'Quand faut-il choisir Veo 3.1 Fast plutot que Veo 3.1 Lite ?',
-            answer:
-              'Choisissez Fast quand vous voulez plus de flexibilite de workflow, un controle audio optionnel et un chemin de montee plus simple vers Veo 3.1. Choisissez Lite quand votre priorite est le testing Veo actuel le moins cher avec audio toujours actif.',
-          },
-        ],
-      },
-    },
-    'kling-3-pro-vs-kling-3-standard': {
-      heroIntro:
-        'Comparez Kling 3 Pro et Kling 3 Standard pour choisir le bon modele Kling IA actuel pour la video multi-shot, les workflows Kling en image-vers-video, les Kling Elements reutilisables et la qualite de sortie avec audio natif.',
-      faq: {
-        title: 'FAQ',
-        subtitle:
-          'Reponses courtes pour choisir entre les tiers Kling Pro et Standard actuels.',
-        items: [
-          {
-            question: 'Quel modele Kling IA actuel convient le mieux a l image-vers-video et aux tests de prompt ?',
-            answer:
-              'Kling 3 Standard convient mieux a des tests de prompt moins couteux et a des tests image-vers-video repetables, tandis que Kling 3 Pro convient mieux quand vous avez besoin d un controle de scene plus serre et de sorties finales plus prioritaires.',
-          },
-          {
-            question: 'Kling 3 Pro et Kling 3 Standard supportent-ils tous les deux les Kling Elements ?',
-            answer:
-              'Oui. Les deux modeles Kling actuels supportent les Kling Elements pour la continuite des personnages et des props, mais Kling 3 Pro reste le meilleur choix quand la sequence est plus exigeante ou que la continuite compte davantage.',
-          },
-          {
-            question: 'Quand faut-il choisir Kling 3 Pro plutot que Kling 3 Standard ?',
-            answer:
-              'Choisissez Kling 3 Pro quand vous avez besoin d un meilleur controle de scene, d une continuite multi-shot plus exigeante et d une sortie finale plus prioritaire. Choisissez Kling 3 Standard quand le controle des couts et les tests repetables comptent davantage.',
-          },
-        ],
-      },
-    },
-  },
-  es: {
-    'seedance-1-5-pro-vs-seedance-2-0': {
-      meta: {
-        title: 'Seedance 1.5 Pro vs Seedance 2.0 | Que cambio, cuando actualizar y mejores objetivos | MaxVideoAI',
-        description:
-          'Compara Seedance 1.5 Pro y Seedance 2.0 en MaxVideoAI para ver que cambia en audio, continuidad multi-shot, referencias, precio y cuando conviene actualizar.',
-      },
-      heroIntro:
-        'Compara Seedance 1.5 Pro y Seedance 2.0 para ver que cambio entre el workflow Seedance Pro anterior y el modelo de video IA de Seedance actual en audio nativo, continuidad multi-shot y workflows con referencias. Usa esta pagina para entender rapido los trade-offs antes de abrir el modelo Seedance actual, la pagina de ejemplos de video IA de Seedance o el workflow de video de Seedance que mejor encaja con tu objetivo.',
-      topCards: [
-        {
-          title: 'Que cambio',
-          body:
-            'Seedance 2.0 es el workflow de video IA de Seedance mas reciente, con mejor continuidad multi-shot, entradas de referencia mas amplias y un camino de produccion audio-first mas actual que Seedance 1.5 Pro.',
-        },
-        {
-          title: 'Cuando quedarse en Seedance 1.5 Pro',
-          body:
-            'Quedate en Seedance 1.5 Pro cuando necesites sobre todo clips cortos y repetibles, setups de camara mas simples y un workflow anterior que ya esta validado en produccion.',
-        },
-        {
-          title: 'Cuando pasar a Seedance 2.0',
-          body:
-            'Actualiza cuando necesites mejor continuidad entre tomas, workflows de audio nativo mas ricos o un modelo actual mas flexible para trabajo creativo de mayor valor.',
-        },
-        {
-          title: 'Mejor para',
-          body:
-            'Usa esta pagina para decidir entre un workflow Seedance anterior pero aun compatible para clips cortos y controlados, y el workflow actual de Seedance 2.0 para anuncios multi-shot, lanzamientos y secuencias mas ambiciosas guiadas por referencias.',
-        },
-      ],
-      primaryLinksTitle: 'Siguientes pasos recomendados',
-      primaryLinks: [
-        {
-          href: '/models/seedance-2-0',
-          label: 'Abrir la pagina del modelo Seedance 2.0',
-        },
-        {
-          href: '/examples/seedance',
-          label: 'Ver ejemplos de video IA de Seedance',
-        },
-        {
-          href: '/models/seedance-2-0-fast',
-          label: 'Abrir la pagina del modelo Seedance 2.0 Fast',
-        },
-        {
-          href: '/ai-video-engines/seedance-2-0-vs-seedance-2-0-fast',
-          label: 'Comparar Seedance 2.0 vs Seedance 2.0 Fast',
-        },
-      ],
-      faq: {
-        title: 'FAQ',
-        subtitle:
-          'Respuestas rapidas para decidir si conviene quedarse en Seedance 1.5 Pro o pasar a Seedance 2.0.',
-        items: [
-          {
-            question: 'Que cambio entre Seedance 1.5 Pro y Seedance 2.0?',
-            answer:
-              'Seedance 2.0 es el modelo mas nuevo, con mejor continuidad multi-shot, workflows de referencia mas amplios y el camino Seedance actual para produccion. Seedance 1.5 Pro sigue siendo util para clips mas cortos, simples y repetibles.',
-          },
-          {
-            question: 'Es Seedance 2.0 mejor que Seedance 1.5 Pro?',
-            answer:
-              'Para la mayoria de workflows actuales, si. Seedance 2.0 es la mejor opcion por defecto si quieres el modelo de video IA de Seedance actual para mas continuidad, flexibilidad y un uso de produccion mas amplio, mientras que Seedance 1.5 Pro sigue siendo util como setup Seedance Pro anterior para clips mas cortos.',
-          },
-          {
-            question: 'Cuando deberia pasar de Seedance 1.5 Pro a Seedance 2.0?',
-            answer:
-              'Actualiza cuando necesites mejor comportamiento multi-shot, workflows de audio nativo mas ricos o mas margen para produccion actual basada en prompts y referencias. Si tu workflow de 1.5 Pro ya es estable para clips cortos, no hace falta mover todos los trabajos de inmediato.',
-          },
-          {
-            question: 'Sigue siendo suficiente Seedance 1.5 Pro para algunos workflows?',
-            answer:
-              'Si. Sigue encajando en clips cinematicos cortos y repetibles, y en equipos que ya tienen patrones de prompt validados en 1.5 Pro y todavia no necesitan el workflow mas amplio de 2.0.',
-          },
-          {
-            question: 'Que modelo es mejor para multi-shot y audio nativo?',
-            answer:
-              'Seedance 2.0 es la mejor opcion para continuidad multi-shot y para el workflow de audio nativo mas actual. Seedance 1.5 Pro conviene tratarlo como una opcion anterior mas simple para clips cortos.',
-          },
-        ],
-      },
-    },
-    'seedance-2-0-vs-seedance-2-0-fast': {
-      heroIntro:
-        'Compara Seedance 2.0 y Seedance 2.0 Fast para elegir el workflow de video IA de Seedance actual mas adecuado segun tu necesidad de trabajo multi-shot final, audio nativo y comparacion de workflow. Usa esta pagina para ver cuando el Seedance estandar encaja mejor en una salida de video de Seedance mas pulida y cuando Fast encaja mejor para pruebas, checks de timing e iteracion mas barata.',
-      primaryLinksTitle: 'Siguientes pasos recomendados',
-      primaryLinks: [
-        {
-          href: '/models/seedance-2-0',
-          label: 'Abrir la pagina del modelo Seedance 2.0',
-        },
-        {
-          href: '/examples/seedance',
-          label: 'Ver ejemplos de video IA de Seedance',
-        },
-        {
-          href: '/models/seedance-2-0-fast',
-          label: 'Abrir la pagina del modelo Seedance 2.0 Fast',
-        },
-      ],
-      faq: {
-        title: 'FAQ',
-        subtitle:
-          'Respuestas cortas para elegir el workflow actual de Seedance adecuado.',
-        items: [
-          {
-            question: 'Que modelo de video IA de Seedance deberia usar para trabajo de borrador?',
-            answer:
-              'Usa Seedance 2.0 Fast para borradores mas baratos, pruebas mas rapidas y comparaciones de workflow. Usa Seedance 2.0 cuando quieras una salida multi-shot mas pulida, audio nativo y un workflow mas listo para produccion.',
-          },
-          {
-            question: 'En que se diferencia Seedance 2.0 de Seedance 2.0 Fast?',
-            answer:
-              'Seedance 2.0 es la opcion actual mas fuerte para trabajo multi-shot pulido, audio nativo y salidas mas exigentes guiadas por referencias, mientras que Seedance 2.0 Fast encaja mejor en borradores mas baratos, checks de timing e iteracion temprana.',
-          },
-          {
-            question: 'Es mejor Seedance 2.0 para una salida de video de Seedance mas pulida?',
-            answer:
-              'Si. Seedance 2.0 encaja mejor cuando el objetivo es una salida de video de Seedance mas pulida, mientras que Fast encaja mejor cuando el objetivo es probar ideas y comparar workflows rapidamente.',
-          },
-        ],
-      },
-    },
-    'veo-3-1-vs-veo-3-1-fast': {
-      heroIntro:
-        'Compara Veo 3.1 y Veo 3.1 Fast para elegir el workflow de video IA Veo 3 actual mas adecuado segun tu necesidad de text-to-video mas pulido, image-to-video, borradores mas rapidos y control sobre el audio nativo.',
-      faq: {
-        title: 'FAQ',
-        subtitle:
-          'Respuestas cortas para elegir el workflow Veo actual adecuado.',
-        items: [
-          {
-            question: 'Como deberia usar Veo 3 para text-to-video y borradores?',
-            answer:
-              'Usa Veo 3.1 Fast para borradores mas baratos, comparacion de prompts text-to-video e iteracion mas rapida. Usa Veo 3.1 cuando quieras una salida final mas fuerte, mejor control guiado por referencias y resultados image-to-video mas pulidos.',
-          },
-          {
-            question: 'Puedo usar tanto Veo 3.1 como Veo 3.1 Fast para image-to-video?',
-            answer:
-              'Si. Ambos pueden manejar workflows de image-to-video, pero Veo 3.1 encaja mejor en resultados mas pulidos, mientras que Veo 3.1 Fast encaja mejor en pruebas mas baratas de prompt y encuadre.',
-          },
-          {
-            question: 'Cuando deberia elegir Veo 3.1 en lugar de Veo 3.1 Fast?',
-            answer:
-              'Elige Veo 3.1 cuando la calidad final, el pulido del audio nativo y un mejor control guiado por referencias importan mas que la velocidad del borrador. Elige Fast cuando el objetivo es una iteracion mas barata y una validacion de workflow mas rapida.',
-          },
-        ],
-      },
-    },
-    'veo-3-1-fast-vs-veo-3-1-lite': {
-      heroIntro:
-        'Compara Veo 3.1 Fast y Veo 3.1 Lite para elegir el workflow de video IA Veo 3 actual mas adecuado segun tu necesidad de borradores text-to-video mas baratos, pruebas image-to-video, comportamiento del audio e iteracion mas rapida.',
-      faq: {
-        title: 'FAQ',
-        subtitle:
-          'Respuestas cortas para elegir entre los tiers Veo Fast y Lite actuales.',
-        items: [
-          {
-            question: 'Que modelo Veo 3 es mejor para pruebas de image-to-video?',
-            answer:
-              'Los dos pueden servir, pero Veo 3.1 Lite encaja mejor en las pruebas image-to-video con audio mas baratas, mientras que Veo 3.1 Fast encaja mejor cuando quieres mas flexibilidad y una subida mas fluida hacia Veo 3.1.',
-          },
-          {
-            question: 'Veo 3.1 Lite o Veo 3.1 Fast es mejor para borradores text-to-video?',
-            answer:
-              'Veo 3.1 Lite encaja mejor cuando quieres los borradores con audio mas baratos. Veo 3.1 Fast encaja mejor cuando quieres mas flexibilidad, audio opcional y un puente mas limpio hacia el workflow principal de Veo 3.1.',
-          },
-          {
-            question: 'Cuando deberia elegir Veo 3.1 Fast en lugar de Veo 3.1 Lite?',
-            answer:
-              'Elige Fast cuando quieras mas flexibilidad de workflow, control opcional del audio y una ruta de subida mas simple hacia Veo 3.1. Elige Lite cuando tu prioridad sea el testing Veo actual mas barato con audio siempre activado.',
-          },
-        ],
-      },
-    },
-    'kling-3-pro-vs-kling-3-standard': {
-      heroIntro:
-        'Compara Kling 3 Pro y Kling 3 Standard para elegir el modelo Kling AI actual adecuado para video multi-shot, workflows Kling de imagen-a-video, Kling Elements reutilizables y calidad de salida con audio nativo.',
-      faq: {
-        title: 'FAQ',
-        subtitle:
-          'Respuestas cortas para elegir entre los tiers actuales de Kling Pro y Standard.',
-        items: [
-          {
-            question: '¿Que modelo Kling AI actual es mejor para imagen-a-video y pruebas de prompt?',
-            answer:
-              'Kling 3 Standard es mejor para pruebas de prompt de menor coste y borradores repetibles de imagen-a-video, mientras que Kling 3 Pro es mejor cuando necesitas un control de escena mas preciso y salidas finales mas prioritarias.',
-          },
-          {
-            question: '¿Kling 3 Pro y Kling 3 Standard soportan ambos los Kling Elements?',
-            answer:
-              'Si. Ambos modelos Kling actuales soportan Kling Elements para continuidad de personajes y props, pero Kling 3 Pro es la mejor opcion cuando la secuencia es mas exigente o la continuidad importa mas.',
-          },
-          {
-            question: '¿Cuando deberia elegir Kling 3 Pro en lugar de Kling 3 Standard?',
-            answer:
-              'Elige Kling 3 Pro cuando necesites mejor control de escena, continuidad multi-shot mas exigente y una salida final mas prioritaria. Elige Kling 3 Standard cuando importe mas controlar el coste y repetir pruebas de borrador.',
-          },
-        ],
-      },
-    },
-  },
+const COMPARE_PAGE_OVERRIDES: ComparePageOverridesByLocale = {
+  en: EN_COMPARE_PAGE_OVERRIDES,
+  fr: FR_COMPARE_PAGE_OVERRIDES,
+  es: ES_COMPARE_PAGE_OVERRIDES,
 };
 
 export function getComparePageOverride(locale: AppLocale, slug: string): ComparePageOverride | undefined {
